text updates
