+++ js/doc-snippets/src/models/imagen.ts
@@ -14,17 +14,29 @@
  * limitations under the License.
  */
 
-import { gemini15Flash, googleAI } from '@genkit-ai/googleai';
+// [START imagen]
+import { imagen3Fast, vertexAI } from '@genkit-ai/vertexai';
+import parseDataURL from 'data-urls';
 import { genkit } from 'genkit';
 
+import { writeFile } from 'node:fs/promises';
+
 const ai = genkit({
-  plugins: [googleAI()],
-  model: gemini15Flash,
+  plugins: [vertexAI({ location: 'us-central1' })],
 });
 
 (async () => {
-  const { text } = await ai.generate(
-    'Invent a menu item for a pirate themed restaurant.'
-  );
-  console.log(text);
+  const { media } = await ai.generate({
+    model: imagen3Fast,
+    prompt: 'photo of a meal fit for a pirate',
+    output: { format: 'media' },
+  });
+
+  if (media === null) throw new Error('No media generated.');
+
+  const data = parseDataURL(media.url);
+  if (data === null) throw new Error('Invalid "data:" URL.');
+
+  await writeFile(`output.${data.mimeType.subtype}`, data.body);
 })();
+// [END imagen]
